fix crash when reloading broken programs
and move the meat of the implementation to the lib folder so that fire can be tested with fire.
